Improve client key handling, expand gitignore
Client ID and Secret are now stored in two categories in the `client_keys.json`
file, and the program fetches the keys from the same file. The `public` key
contains the default client keys, and the user can add their own client keys
under the `personal` key in the JSON file. Also, added a HUGE bunch of file
formats to the `.gitignore` file